requirements.txt
# Core FastAPI and Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-settings==2.1.0

# Google APIs
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# OpenAI and AI Models
openai==1.6.1
tiktoken==0.5.2

# Vector Database and Embeddings
qdrant-client==1.7.0
numpy==1.24.3
scipy==1.11.4

# Text Processing and NLP
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.7.2

# Async and HTTP
httpx==0.25.2
aiofiles==23.2.1
asyncio-throttle==1.0.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2
